+++ contracts/libraries/calculate.sol
@@ -8,7 +8,11 @@ library calculate {
 
     /// Pre: value denominated in [base unit of melonAsset]
     /// Post: Share price denominated in [base unit of melonAsset * base unit of share / base unit of share] == [base unit of melonAsset]
-    function pricePerShare(uint256 value, uint256 baseUnitsPerShare, uint256 totalSupply)
+    function pricePerShare(
+        uint256 value,
+        uint256 baseUnitsPerShare,
+        uint256 totalSupply
+    )
         constant
         returns (uint256)
     {
@@ -18,7 +22,12 @@ library calculate {
 
     /// Pre: numShares denominated in [base unit of melonAsset], baseUnitsPerShare not zero
     /// Post: priceInRef denominated in [base unit of melonAsset]
-    function priceForNumShares(uint256 numShares, uint256 baseUnitsPerShare, uint256 nav, uint256 totalSupply)
+    function priceForNumShares(
+        uint256 numShares,
+        uint256 baseUnitsPerShare,
+        uint256 nav,
+        uint256 totalSupply
+    )
         constant
         returns (uint256)
     {
@@ -28,7 +37,14 @@ library calculate {
 
     /// Pre: numShares denominated in [base unit of melonAsset], baseUnitsPerShare not zero
     /// Post: priceInRef denominated in [base unit of melonAsset]
-    function subscribePriceForNumShares(uint256 numShares, uint256 subscriptionFee, uint256 baseUnitsPerShare, uint256 feeDivisor, uint256 nav, uint256 totalSupply)
+    function subscribePriceForNumShares(
+        uint256 numShares,
+        uint256 baseUnitsPerShare,
+        uint256 subscriptionFee,
+        uint256 feeDivisor,
+        uint256 nav,
+        uint256 totalSupply
+    )
         constant
         returns (uint256)
     {
@@ -39,7 +55,11 @@ library calculate {
 
     /// Pre: Decimals in assets must be equal to decimals in PriceFeed for all entries in Universe
     /// Post: Gross asset value denominated in [base unit of melonAsset]
-    function grossAssetValue(uint256[] storage holdings, uint256[] storage prices, uint256[] storage decimals)
+    function grossAssetValue(
+        uint256[] storage holdings,
+        uint256[] storage prices,
+        uint256[] storage decimals
+    )
         constant
         returns (uint256 result)
     {
@@ -50,14 +70,28 @@ library calculate {
 
     /// Pre: Gross asset value and sum of all applicable and unclaimed fees has been calculated
     /// Post: Net asset value denominated in [base unit of melonAsset]
-    function netAssetValue(uint256 gav, uint256 rewardsUnclaimed) constant returns (uint256) { return gav.sub(rewardsUnclaimed); }
+    function netAssetValue(
+        uint256 gav,
+        uint256 rewardsUnclaimed
+    )
+        constant
+        returns (uint256)
+    {
+        return gav.sub(rewardsUnclaimed);
+    }
 
     /* Function invariant
      *  for timeDifference == 0 => returns 0
      */
     /// Post: Reward denominated in referenceAsset
-    function managementReward(uint managementRewardRate, uint timeDifference, uint gav, uint divisorFee)
-        constant returns (uint)
+    function managementReward(
+        uint managementRewardRate,
+        uint timeDifference,
+        uint gav,
+        uint divisorFee
+    )
+        constant
+        returns (uint)
     {
         uint absoluteChange = timeDifference * gav;
         return absoluteChange * managementRewardRate / divisorFee;
@@ -67,8 +101,14 @@ library calculate {
      *  for deltaDifference == 0 => returns 0
      */
     /// Post: Reward denominated in referenceAsset
-    function performanceReward(uint performanceRewardRate, uint sharePriceDifference, uint totalSupply, uint divisorFee)
-        constant returns (uint)
+    function performanceReward(
+        uint performanceRewardRate,
+        uint sharePriceDifference,
+        uint totalSupply,
+        uint divisorFee
+    )
+        constant
+        returns (uint)
     {
         if (sharePriceDifference <= 0) return 0;
         uint absoluteChange = sharePriceDifference * totalSupply;
@@ -77,9 +117,22 @@ library calculate {
 
     /// Pre: Gross asset value has been calculated
     /// Post: The sum and its individual parts of all applicable fees denominated in [base unit of melonAsset]
-    function rewards(uint256 lastPayoutTime, uint256 managementRewardRate, uint256 performanceRewardRate, uint256 gav, uint256 lastSharePrice, uint256 totalSupply, uint256 baseUnitsPerShare, uint256 divisorFee)
+    function rewards(
+        uint256 lastPayoutTime,
+        uint256 managementRewardRate,
+        uint256 performanceRewardRate,
+        uint256 gav,
+        uint256 lastSharePrice,
+        uint256 totalSupply,
+        uint256 baseUnitsPerShare,
+        uint256 divisorFee
+    )
         constant
-        returns (uint256 management, uint256 performance, uint256 unclaimed)
+        returns (
+            uint256 management,
+            uint256 performance,
+            uint256 unclaimed
+        )
     {
         uint256 timeDifference = now.sub(lastPayoutTime);
         management = managementReward(managementRewardRate, timeDifference, gav, divisorFee);